moved default live effect back into to_ass.py
